refactor(backend): refactor resource helpers module structure
