+++ src-tauri/src/resource/helpers/mod_loader_meta/forge.rs
@@ -1,4 +1,4 @@
-use super::misc::get_download_api;
+use super::super::misc::get_download_api;
 use crate::error::{SJMCLError, SJMCLResult};
 use crate::instance::models::misc::ModLoaderType;
 use crate::resource::models::{ModLoaderResourceInfo, ResourceError, ResourceType, SourceType};
