+++ src-tauri/src/resource/helpers/mod_loader_meta/mod.rs
@@ -1,0 +1,3 @@
+pub mod fabric;
+pub mod forge;
+pub mod neoforge;
