+++ src-tauri/src/resource/helpers/mod_loader_meta/neoforge.rs
@@ -1,4 +1,4 @@
-use super::misc::get_download_api;
+use super::super::misc::get_download_api;
 use crate::error::SJMCLResult;
 use crate::instance::models::misc::ModLoaderType;
 use crate::resource::models::{ModLoaderResourceInfo, ResourceError, ResourceType, SourceType};
